requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-telegram-bot[webhooks]==21.10
sqlmodel==0.0.22
pydantic-settings==2.7.1
python-dotenv==1.0.1
